Remove files/objects.
